Extract prescan result type
