+++ src/lowering/prescanTypes.ts
@@ -1,0 +1,15 @@
+import type { EaExprNode, OpDeclNode, TypeExprNode, VarDeclNode } from '../frontend/ast.js';
+import type { Callable } from './loweringTypes.js';
+
+export interface PrescanResult {
+  localCallablesByFile: Map<string, Map<string, Callable>>;
+  visibleCallables: Map<string, Callable>;
+  localOpsByFile: Map<string, Map<string, OpDeclNode[]>>;
+  visibleOpsByName: Map<string, OpDeclNode[]>;
+  declaredOpNames: Set<string>;
+  declaredBinNames: Set<string>;
+  storageTypes: Map<string, TypeExprNode>;
+  moduleAliasTargets: Map<string, EaExprNode>;
+  moduleAliasDecls: Map<string, VarDeclNode>;
+  rawAddressSymbols: Set<string>;
+}
